Add package upgrade at system startup and periodically
Create a system service, which polls the smart repositories for
system updates.  At runtime, configuration can be specified in
the file /etc/sysupdate/config.local to override the default
configuration specified in /etc/sysupdate/config.default.

By default, the update occurs slightly more frequently than once
per week.

The initial implementation:

- Does allow configuration of the behavior, among three choices:

	* not run (default)
	* check for updates, and notify via /etc/motd
	* auto install : download and install everything without
	  user interaction

    Note that another option, cron-driven, is possible.  This
	can be done by manually disabling the service and manually
	adding a cron job.

- Does allow runtime configuration of the polling period.

- Does allow the configuration directory to be changed.  However,
if this is done, it must be done in the sysupdate config file
and the systemd config file if systemd is used, to be effective.

- Implements both systemd and sysvinit installation recipes,
but only systemd was significantly tested.

- Does log to a file, but does not handle log rotation.

- Does not install the service automatically, unless configured
in local.conf at build time.  The docs need to be updated with
instructions on how to do this.

- Does not make any attempt to syncronize the poll to a particular
time of day.  If this is desired, the updates can be done from
cron.  The docs need to be updated with instructions on how to
do this.

- Does not make any attempt to correct for skew caused by the
duration of the poll itself.

- Does not check whether smartpm has been configured with valid
channels.

It is expected that the deficiencies among the points listed
above will be cleaned up at an appropriate future time.

Signed-off-by: T.O. Radzy Radzykewycz <radzy@windriver.com>